# Add simd sha256 intrinsics for x86 machines