@RCasatta 2023-07-27
Thanks for this!

I confirm I have similar benchmark improvements on my machine.

The error in CI is just formatting, a `cargo +nightly fmt` is needed.

Any rationale on choosing the code to base the implementation on over using the code in bitcoin core or in sha2 crate?

Is having some fuzzing between hardware and software implementation a good idea?