@apoelstra 2023-07-28
A private function that is only called once is pretty-much guaranteed to be inlined.