@sanket1729 2023-08-02
>  we are very often taking 2/4/8/etc neighboring blocks of 64 bytes and hashing each block at once. IIRC in Bitcoin Core pre-sha-ni it made a huge difference

I did some superficial `simd` instruction counting for 2/4/8 blocks interleaving methods from bitcoin core. 
Using `Tranform2_sha256d` from bitcoin core saves about 28 calls to `_mm_sha256msg1_epu32` compared to calling sha256::hash repeatedly. Both of the still have the same number of calls to `_mm_sha256rnds2_epu32`.

Overall, one sha256d(3 transforms) computation done without transform2 takes about 192 _mm_sha256rnds2_epu32 and 84 _mm_sha256msg1_epu32. 

Based on my superficial evaluation, there is still some performance benefit to be gained, but certainly not a significant one.