@tcharding 2023-08-02
Oh yes , my bad, I should have convinced myself of unit test coverage before posting.