@tcharding 2023-08-02
What's the plan from here @sanket1729, are we going to do the other hashes from noloader code too (sha1, sha512)?