@sanket1729 2023-08-04
Ah thanks for elaborating. Makes much more sense now.

On Thu, Aug 3, 2023, 11:45 PM Matt Corallo ***@***.***> wrote:

> I did some superficial simd instruction counting for 2/4/8 blocks
> interleaving methods from bitcoin core.
>
> So my point about performance here wasn't about the instruction count, but
> rather the data inverleaving allowing the CPU to better parallelize
> internally, which can allow for much higher throughput without fewer
> instructions.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/rust-bitcoin/rust-bitcoin/pull/1962#issuecomment-1665094878>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ABUQEOPPDULLNXLOFCWQM63XTSLAHANCNFSM6AAAAAA2Z3LIJ4>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>