@apoelstra 2023-08-16
Oh, I don't think I noticed your ACK @tcharding. Can you re-ack with the commit ID?